requirements.txt
asgiref==3.6.0
black==23.1.0
cfgv==3.3.1
click==8.1.3
coverage==6.1
distlib==0.3.6
Django==4.1.7
django-filter==21.1
django-fsm==2.8.1
djangorestframework==3.14.0
filelock==3.9.0
flake8==4.0.1
identify==2.5.19
mccabe==0.6.1
mypy-extensions==1.0.0
nodeenv==1.7.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.1.1
pre-commit==2.15.0
psycopg2==2.9.5
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
toml==0.10.2
virtualenv==20.20.0
